This PR updates ComplEx to use the new-style `ERModel`

### Choosing a compute kernel for ComplEx

<summary>Benchmark on RTX2080 Ti</summary>
<details>

```python
from typing import Tuple

import torch
from torch.utils.benchmark import Compare, Timer
from tqdm.auto import tqdm

import pykeen.nn.compute_kernel as cc


def prepare(b: int, n: int, d: int, use_case: str, device: str) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
    hs = (b, 1, 1, 1, 2 * d)
    rs = (b, 1, 1, 1, 2 * d)
    ts = (b, 1, 1, 1, 2 * d)
    if use_case == "score_hrt":
        pass
    elif use_case == "score_t":
        ts = (1, 1, 1, n, 2 * d)
    elif use_case == "score_h":
        hs = (1, n, 1, 1, 2 * d)
    else:
        raise ValueError(use_case)

    return (
        torch.rand(*hs, device=device),
        torch.rand(*rs, device=device),
        torch.rand(*ts, device=device),
    )


def main():
    if torch.cuda.is_available():
        devices = ["cuda"]
    else:
        devices = ["cpu"]
    print(devices)
    methods = (
        cc._complex_broadcast_optimized,
        cc._complex_direct,
        cc._complex_einsum,
        cc._complex_native_complex,
        cc._complex_native_complex_select,
        cc._complex_select,
        cc._complex_stacked,
        cc._complex_stacked_select,
    )
    settings = [
        (256, 15_000, 512, "score_hrt"),
        (32, 15_000, 512, "score_h"),
        (32, 15_000, 512, "score_t"),
        (256, 40_000, 512, "score_hrt"),
        (8, 40_000, 512, "score_h"),
        (8, 40_000, 512, "score_t"),
    ]
    total = len(devices) * len(methods) * len(settings)
    compare = Compare(
        results=list(
            tqdm(
                (
                    Timer(
                        stmt=f"cc.{method.__name__}(h=h, r=r, t=t)",
                        setup=f'h, r, t = prepare(b={b}, n={n}, d={d}, use_case="{use_case}", device="{device}")',
                        globals=dict(
                            cc=cc,
                            prepare=prepare,
                        ),
                        env=f"b={b}, n={n}, d={d}",
                        description=use_case,
                    ).blocked_autorange()
                    for b, n, d, use_case in settings
                    for device in devices
                    for method in methods
                ),
                total=total,
            )
        )
    )
    compare.colorize()
    compare.trim_significant_figures()
    compare.print()


if __name__ == "__main__":
    main()

```

#### Results
```
                                                                             |  score_hrt  |  score_h  |  score_t
1 threads: ------------------------------------------------------------------------------------------------------
  (b=256, n=15000, d=512)  cc._complex_broadcast_optimized(h=h, r=r, t=t)    |     221     |           |         
                           cc._complex_direct(h=h, r=r, t=t)                 |     140     |           |         
                           cc._complex_einsum(h=h, r=r, t=t)                 |     423     |           |         
                           cc._complex_native_complex(h=h, r=r, t=t)         |      71     |           |         
                           cc._complex_native_complex_select(h=h, r=r, t=t)  |     100     |           |         
                           cc._complex_select(h=h, r=r, t=t)                 |     170     |           |         
                           cc._complex_stacked(h=h, r=r, t=t)                |     140     |           |         
                           cc._complex_stacked_select(h=h, r=r, t=t)         |     190     |           |         
  (b=256, n=40000, d=512)  cc._complex_broadcast_optimized(h=h, r=r, t=t)    |     224     |           |         
                           cc._complex_direct(h=h, r=r, t=t)                 |     140     |           |         
                           cc._complex_einsum(h=h, r=r, t=t)                 |     423     |           |         
                           cc._complex_native_complex(h=h, r=r, t=t)         |      75     |           |         
                           cc._complex_native_complex_select(h=h, r=r, t=t)  |     130     |           |         
                           cc._complex_select(h=h, r=r, t=t)                 |     170     |           |         
                           cc._complex_stacked(h=h, r=r, t=t)                |     140     |           |         
                           cc._complex_stacked_select(h=h, r=r, t=t)         |     190     |           |         
  (b=32, n=15000, d=512)   cc._complex_broadcast_optimized(h=h, r=r, t=t)    |             |   21720   |   21710 
                           cc._complex_direct(h=h, r=r, t=t)                 |             |   36450   |   21670 
                           cc._complex_einsum(h=h, r=r, t=t)                 |             |   90300   |     336 
                           cc._complex_native_complex(h=h, r=r, t=t)         |             |   10900   |   11100 
                           cc._complex_native_complex_select(h=h, r=r, t=t)  |             |   11100   |   11400 
                           cc._complex_select(h=h, r=r, t=t)                 |             |   20940   |    3445 
                           cc._complex_stacked(h=h, r=r, t=t)                |             |   36890   |   22290 
                           cc._complex_stacked_select(h=h, r=r, t=t)         |             |   40140   |    7480 
  (b=8, n=40000, d=512)    cc._complex_broadcast_optimized(h=h, r=r, t=t)    |             |   14500   |   14480 
                           cc._complex_direct(h=h, r=r, t=t)                 |             |   24290   |   14450 
                           cc._complex_einsum(h=h, r=r, t=t)                 |             |   69690   |     524 
                           cc._complex_native_complex(h=h, r=r, t=t)         |             |    7260   |    7870 
                           cc._complex_native_complex_select(h=h, r=r, t=t)  |             |    7895   |    8510 
                           cc._complex_select(h=h, r=r, t=t)                 |             |   14000   |    2302 
                           cc._complex_stacked(h=h, r=r, t=t)                |             |   25510   |   16270 
                           cc._complex_stacked_select(h=h, r=r, t=t)         |             |   27780   |    6411 

Times are in microseconds (us).
```
</details>

TL;DR: native complex is fastest for `score_hrt` / `score_t`. For `score_h` einsum is faster.


### Comparison to `master`

<summary>Benchmark on RTX2080 Ti</summary>
<details>

Using default settings, and (a somewhat arbitrary) fixed `batch_size=256` on RTX 2080 Ti: `nan` indicates OOM.

#### old
| model   | dataset   |   training |   evaluation |   score_hrt |       score_h |       score_t |  device   | git_hash   |                                                                 
|:--------|:----------|-----------:|-------------:|------------:|--------------:|--------------:|:-----------|:-----------|
| complex | fb15k237  |   6.8168   |    16.3357   | 0.000343446 | nan           |   0.0666306   |  cuda     | b3375345   |
| complex | nations   |   0.182782 |     0.127749 | 0.000350535 |   0.000319821 |   0.000316372 |  cuda     | b3375345   |
| complex | wn18rr    |   3.01705  |     6.71486  | 0.000342343 | nan           | nan           |  cuda     | b3375345   |

#### new
| model   | dataset   |   training |   evaluation |   score_hrt |       score_h |       score_t |  device   | git_hash   |                                                                 
|:--------|:----------|-----------:|-------------:|------------:|--------------:|--------------:|:-----------|:-----------|
| complex | fb15k237  |   4.96571  |     7.2699   | 0.000283919 |   0.0332715   |   0.0335903   | cuda     | e03bba90   |
| complex | nations   |   0.154651 |     0.127478 | 0.000278247 |   0.000257923 |   0.000254081 |  cuda     | e03bba90   |
| complex | wn18rr    |   2.57455  |     3.09619  | 0.000286655 | nan           | nan           |  cuda     | e03bba90   |
</details>

TL;DR: new version seems to be faster, and sometimes more memory efficient.