@mberr 2021-11-15
Using default settings, and (a somewhat arbitrary) fixed `batch_size=256` on RTX 2080 Ti: `nan` indicates OOM.

#### old
| model   | dataset   |   training |   evaluation |   score_hrt |       score_h |       score_t |  device   | git_hash   |                                                                 
|:--------|:----------|-----------:|-------------:|------------:|--------------:|--------------:|:-----------|:-----------|
| complex | fb15k237  |   6.8168   |    16.3357   | 0.000343446 | nan           |   0.0666306   |  cuda     | b3375345   |
| complex | nations   |   0.182782 |     0.127749 | 0.000350535 |   0.000319821 |   0.000316372 |  cuda     | b3375345   |
| complex | wn18rr    |   3.01705  |     6.71486  | 0.000342343 | nan           | nan           |  cuda     | b3375345   |

#### new
| model   | dataset   |   training |   evaluation |   score_hrt |       score_h |       score_t |  device   | git_hash   |                                                                 
|:--------|:----------|-----------:|-------------:|------------:|--------------:|--------------:|:-----------|:-----------|
| complex | fb15k237  |   4.96571  |     7.2699   | 0.000283919 |   0.0332715   |   0.0335903   | cuda     | e03bba90   |
| complex | nations   |   0.154651 |     0.127478 | 0.000278247 |   0.000257923 |   0.000254081 |  cuda     | e03bba90   |
| complex | wn18rr    |   2.57455  |     3.09619  | 0.000286655 | nan           | nan           |  cuda     | e03bba90   |