@cthoyt 2021-11-15
2 things before finishing: can you add unit tests for the new utility functions, and can you add a bit more prose about what the differences between the things you tested were and why you think some performed better?