@mberr 2021-11-16
> 2 things before finishing: can you add unit tests for the new utility functions, and can you add a bit more prose about what the differences between the things you tested were and why you think some performed better?

As discussed in PM, this information is in the updated issue description (hidden in collapsibles :smile: )